creating Tung Tung Tung Sahur PowerUp
i added PowerUp Manager, there you can change the duration of the power up and add sound to it.
then PowerUpSpawner is for controlling the interval and chance of the power up spawning.
and PlayerPowerUps is for all the sprite and animation transformation

*note*
both the animation and animator for rengi and tung tung tung can be fixed a little bit (and attack animation for tung tung not working for this moment)